Cleanup and refactor CDK classes subdir, SCRUM-2204
