+++ cdk_infra/cdk_classes/ecs_stack.py
