+++ cdk_infra/cdk_classes/lambda_stack.py
@@ -34,7 +34,7 @@ class LambdaEcsTrigger:
 		# Copy helper file into lambda bundle
 		dirname = os.path.dirname(os.path.realpath(__file__))
 		shutil.copyfile(os.path.join(dirname, '..','..','app','interfaces','helper.py'),
-		                os.path.join(dirname, 'lambda_bundle', 'helper.py'))
+		                os.path.join(dirname, '..', 'lambda_bundle', 'helper.py'))
 
 		# Create lambda function
 		aws_lambda_fn = aws_lambda.Function(scope, "agrDbBackupsLambdaTrigger",
@@ -42,7 +42,7 @@ class LambdaEcsTrigger:
 			description='Lambda function to trigger a backup or restore of a postgres databases to or from S3, through ECS',
 			runtime=aws_lambda.Runtime.PYTHON_3_7,
 			handler="ecs_trigger.lambda_handler",
-			code=aws_lambda.Code.from_asset(os.path.join(dirname, "lambda_bundle")),
+			code=aws_lambda.Code.from_asset(os.path.join(dirname, '..', 'lambda_bundle')),
 			environment={
 				'AGRDB_ECS_CLUSTER': ecs_cluster_arn,
 				'AGRDB_ECS_TASK_DEF': ecs_task_def_arn,
@@ -53,10 +53,10 @@ class LambdaEcsTrigger:
 			role=excecution_role,
 			timeout=Duration.seconds(60))
 
-		os.remove(os.path.join(dirname, 'lambda_bundle', 'helper.py'))
+		os.remove(os.path.join(dirname, '..', 'lambda_bundle', 'helper.py'))
 
 		# Add targets to nightly backup event rule for every DB requiring nightly backup
-		nightly_backup_targets = json.load(open(os.path.join(dirname, 'resources', 'backup_list.json'), 'r'))
+		nightly_backup_targets = json.load(open(os.path.join(dirname, '..', 'resources', 'backup_list.json'), 'r'))
 
 		for event_obj in nightly_backup_targets:
 			nightly_backups_event_rule.add_target(aws_events_targets.LambdaFunction(
