+++ cdk_infra/lambda_bundle/ecs_trigger.py
