+++ cdk_infra/tests/unit/test_cdk_infra_stack.py
@@ -1,10 +1,10 @@
 import aws_cdk as core
 import aws_cdk.assertions as assertions
 
-from cdk_infra.cdk_infra_stack import CdkInfraStack
+from cdk_classes.cdk_infra_stack import CdkInfraStack
 
 # example tests. To run these tests, uncomment this file along with the example
-# resource in cdk_infra/cdk_infra_stack.py
+# resource in cdk_classes/cdk_infra_stack.py
 def test_sqs_queue_created():
 	app = core.App()
 	stack = CdkInfraStack(app, "cdk-infra")
